@@ -16,8 +16,6 @@ setup(
     install_requires=[
         "charset-normalizer >= 2.0.0",
         "cryptography >= 36.0.0",
-        'typing_extensions; python_version < "3.8"',
-        'importlib_metadata; python_version < "3.8"',
     ],
     extras_require={
         "dev": ["pytest", "nox", "black", "mypy == 0.931"],
@@ -41,7 +39,7 @@ setup(
         "layout analysis",
         "text mining",
     ],
-    python_requires=">=3.6",
+    python_requires=">=3.8",
     classifiers=[
         "Programming Language :: Python",
         "Programming Language :: Python :: 3.8",
