@@ -8,7 +8,7 @@ from pdfminer.pdftypes import PDFObjectNotFound, dict_value, int_value
 from tests.helpers import absolute_sample_path
 
 
-class TestPdfDocument(object):
+class TestPdfDocument:
     def test_get_zero_objid_raises_pdfobjectnotfound(self):
         with open(absolute_sample_path("simple1.pdf"), "rb") as in_file:
             parser = PDFParser(in_file)
